Move cleanse_metadata() into reproducibly.py
So that reproducibly can be distributed as a single file script.
