+++ cleanse_metadata.py
@@ -15,25 +15,10 @@ https://github.com/NixOS/nixpkgs/blob/master/pkgs/development/python-modules/set
 # cleanse_metadata.py
 # Copyright 2023 Keith Maxwell
 # SPDX-License-Identifier: MPL-2.0
-import gzip
-import tarfile
 from argparse import ArgumentParser
-from datetime import datetime
-from os import utime
 from pathlib import Path
-from shutil import copyfileobj
-from stat import S_IWGRP
-from stat import S_IWOTH
-from tempfile import TemporaryDirectory
 
-# - Built distributions are created from source distributions
-# - Source distributions are typically gzipped tar files
-# - Built distributions are typically zip files
-# - The default date for this script is the earliest date supported by both
-# - The minimum date value supported by zip files, is documented in
-#   <https://github.com/python/cpython/blob/3.11/Lib/zipfile.py>.
-EARLIEST_DATE = datetime(1980, 1, 1, 0, 0, 0).timestamp()
-
+from reproducibly import cleanse_metadata
 
 # [[[cog
 # import cog
@@ -42,40 +27,6 @@ EARLIEST_DATE = datetime(1980, 1, 1, 0, 0, 0).timestamp()
 # ]]]
 __version__ = "0.0.1.dev1"
 # [[[end]]]
-
-
-def cleanse_metadata(path_: Path, mtime: float = EARLIEST_DATE) -> int:
-    """Cleanse metadata from a single source distribution"""
-    path = path_.absolute()
-
-    mtime = max(mtime, EARLIEST_DATE)
-
-    with TemporaryDirectory() as directory:
-        with tarfile.open(path) as tar:
-            tar.extractall(path=directory)
-
-        path.unlink(missing_ok=True)
-        (extracted,) = Path(directory).iterdir()
-        uncompressed = f"{extracted}.tar"
-
-        prefix = directory.removeprefix("/") + "/"
-
-        def filter_(tarinfo: tarfile.TarInfo) -> tarfile.TarInfo:
-            tarinfo.mtime = int(mtime)
-            tarinfo.uid = tarinfo.gid = 0
-            tarinfo.uname = tarinfo.gname = "root"
-            tarinfo.mode = tarinfo.mode & ~S_IWGRP & ~S_IWOTH
-            tarinfo.path = tarinfo.path.removeprefix(prefix)
-            return tarinfo
-
-        with tarfile.open(uncompressed, "w") as tar:
-            tar.add(extracted, filter=filter_)
-
-        with gzip.GzipFile(filename=path, mode="wb", mtime=mtime) as file:
-            with open(uncompressed, "rb") as tar:
-                copyfileobj(tar, file)
-        utime(path, (mtime, mtime))
-    return 0
 
 
 def parse_args(args: list[str] | None):
